+++ src/main/java/entities/todo/CommonToDoFactory.java
@@ -9,7 +9,7 @@ public class CommonToDoFactory {
      * @param assignedTo people who are assigned to with this to_do.
      * @return a CommonToDo class.
      */
-    public static CommonToDo create(Integer ID, String target, String[] assignedTo, String progress) {
+    public static CommonToDo create(long ID, String target, String[] assignedTo, String progress) {
         return new CommonToDo(ID, target, assignedTo, progress);
     }
